requirements.txt
anthropic
python-dotenv
pydantic
bs4
requests